+++ cases/testcases/refer_clojure_exclude/red.clj
@@ -1,0 +1,10 @@
+(ns testcases.refer-clojure-exclude.red
+  (:require
+   [clojure.test :refer [deftest is]]))
+
+(defn sut [x]
+  (* x 2))
+
+(deftest uses-update
+  (update {} :f inc)
+  (is (= 42 (sut 21))))