+++ changes.md
@@ -9,6 +9,8 @@
 
 #### Bugfixes
 
+* Handle `:refer-clojure :exclude` better in `:suspicious-test`.
+  * Closes https://github.com/jonase/eastwood/issues/185
 * Exclude Lein checkouts' directories from the inferred `:source-paths` (source path inference being Eastwood's fallback to a lack of explicitly specified `:source-paths` / `:test-paths`). 
 
 ## Changes from 0.9.6 to 0.9.7